97 add clang support (#98)
* change: Fixed up various clang issues, updated spdlog to v1.17.0 and modified build solution to use clang-cl

* change: Fixed up various clang warnings

* change: additional clang fixes

* change: various clang fixes, added ability to search for assembly files given root folder, updated Coral submodule

* change: Add various fixes for Linux compilation

+++ Razor/src/Razor/Systems/RSRenderPass.cpp
@@ -70,14 +70,14 @@ namespace Razor
 		if (data.mDebugLineProperties.size() > 0) {
 			Renderer->UseShader(defaultShader->ID);
 			HandleProperties(data.mDebugLineProperties[0].GetPropertySlot(0), defaultShader);
-			for (int i = 0; i < data.mDebugLines.size(); i++)
+			for (size_t i = 0; i < data.mDebugLines.size(); i++)
 			{
 				Renderer->DrawLine(data.mDebugLines[i]);
 			}
 		}
 		data.mDebugLineProperties.clear();
 		
-		for (int i = 0; i < data.mDebugTriangles.size(); i++)
+		for (size_t i = 0; i < data.mDebugTriangles.size(); i++)
 		{
 			Renderer->UseShader(defaultShader->ID);
 			HandleProperties(data.mDebugTriangleProperties[i].GetPropertySlot(0), defaultShader);
